backend/requirements.txt
slowapi
# Web Framework
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# RAG & LLM Orchestration
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.11

# Vector Store & DB (Supabase)
supabase>=2.4.0

# Git & File Handling
GitPython>=3.1.41
tiktoken>=0.5.2

# Google Generative AI client
google-generativeai>=0.8.0

# Utilities
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.1

# Testing
pytest>=8.0.0